move get_confirmation to utility
